Move test/integration/framework to test/e2e/framework/managed
