+++ test/e2e/framework/managed/cluster.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package framework
+package managed
 
 import (
 	"time"
